Always perform wrapping_adds
All numbers the LC-3 computer handles are unsigned. In general
it does not care about the sign of a number. The only exception
is when setting the condition flag register.

Negative numbers in LC-3 are 2's complement so we can add them up
correctly by wrapping and ignoring any carry bits when adding up the
most significant bits. We can achieve this by using wrapping_add
everywhere.