Support memory mapped keyboard registers
I now wrap every memory read in read_mem to intercept reads to
MRKBSR. To make sure the terminal never buffers STDIN lines I
instantiate Getch once when Computer starts. The Getch constructor
will handle this. I keep it in scope as long as Computer is running.

To poll for keyboard input when MRKBSR is read I had to resort to libc
because all reading of streams in Rust is blocking. With the libc
bindings I could directly call select.

With these changes 2048 works!